Configure MongoDB and add database restore scripts

+++ utilities/run_pipeline.ps1
@@ -1,3 +1,14 @@
+$python = Join-Path $PSScriptRoot "..\.venv\Scripts\python.exe"
+
+if (-not (Test-Path $python)) {
+    Write-Host "Virtual environment not found."
+    Write-Host "Run:"
+    Write-Host "    python -m venv .venv"
+    Write-Host "    .\.venv\Scripts\Activate.ps1"
+    Write-Host "    python -m pip install -r requirements.txt"
+    exit 1
+}
+
 "Pipeline started $(Get-Date)" |
 Out-File "logs\startup_check.log" -Append
 
@@ -6,7 +17,7 @@ Write-Host "DIVIDEND TRACKER PIPELINE"
 Write-Host "Started: $(Get-Date)"
 Write-Host "==================================="
 
-Set-Location "C:\Users\pauls\Documents\dev\python\proj\Dividend-Tracker-2-0"
+Set-Location "C:\Projects\dividend-tracker"
 
 #Clean up logs\run_pipeline_$stamp
 $logDir = ".\logs"
@@ -21,7 +32,7 @@ Get-ChildItem $logDir -File |
 $stamp = Get-Date -Format "yyyyMMdd_HHmmss"
 Start-Transcript -Path ".\logs\run_pipeline_$stamp.log"
 
-python -m utilities.health_check
+& $python -m utilities.health_check
 
 if ($LASTEXITCODE -ne 0)
 {
@@ -43,32 +54,32 @@ Write-Host "UTC Bucket : $utcBucket"
 switch ($day)
 {
     "Monday" {
-        python -m collectors.save_daily_metrics
+       & $python -m collectors.save_daily_metrics
     }
 
     "Tuesday" {
-        python -m collectors.save_daily_metrics
+       & $python -m collectors.save_daily_metrics
     }
 
     "Wednesday" {
-        python -m collectors.save_daily_metrics
+       & $python -m collectors.save_daily_metrics
     }
 
     "Thursday" {
-        python -m collectors.save_daily_metrics
+       & $python -m collectors.save_daily_metrics
     }
 
     "Friday" {
-        python -m collectors.save_dividend_history
+       & $python -m collectors.save_dividend_history
     }
 
     "Saturday" {
-        python -m collectors.save_free_cash_flow
+       & $python  -m collectors.save_free_cash_flow
     }
 
     "Sunday" {
-        python -m analyzers.dividend_safety_analyzer
-        python -m analyzers.portfolio_scorecard
+       & $python -m analyzers.dividend_safety_analyzer
+       & $python -m analyzers.portfolio_scorecard
     }
 }
 Stop-Transcript
